# spec/swagger/v1/swagger.yaml
openapi: 3.0.1
info:
  title: Crowd Parlay Social API
  description: RESTful API of the Crowd Parlay's Social service.
  version: '1.0'
paths:
  '/api/v1/authors/{authorId}':
    get:
      tags:
        - Authors
      summary: Returns author with the specified ID.
      parameters:
        - name: authorId
          in: path
          required: true
          style: simple
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/AuthorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/AuthorDto'
  '/api/v1/comments/{commentId}':
    get:
      tags:
        - Comments
      summary: Returns comment with the specified ID.
      parameters:
        - name: commentId
          in: path
          required: true
          style: simple
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/CommentDto'
            application/json:
              schema:
                $ref: '#/components/schemas/CommentDto'
            text/json:
              schema:
                $ref: '#/components/schemas/CommentDto'
  /api/v1/comments:
    get:
      tags:
        - Comments
      summary: Returns all comments created by author with the specified ID.
      parameters:
        - name: authorId
          in: query
          style: form
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          style: form
          schema:
            type: integer
            format: int32
        - name: size
          in: query
          style: form
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommentDto'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommentDto'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommentDto'
    post:
      tags:
        - Comments
      summary: Creates a top-level comment in discussion.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/CommentRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/CommentRequest'
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/CommentDto'
            application/json:
              schema:
                $ref: '#/components/schemas/CommentDto'
            text/json:
              schema:
                $ref: '#/components/schemas/CommentDto'
  '/api/v1/comments/{targetCommentId}':
    post:
      tags:
        - Comments
      summary: Creates a reply to comment with the specified ID.
      parameters:
        - name: targetCommentId
          in: path
          required: true
          style: simple
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReplyRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/ReplyRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/ReplyRequest'
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/CommentDto'
            application/json:
              schema:
                $ref: '#/components/schemas/CommentDto'
            text/json:
              schema:
                $ref: '#/components/schemas/CommentDto'
  '/api/v1/discussions/{discussionId}':
    get:
      tags:
        - Discussions
      summary: Returns discussion with the specified ID.
      parameters:
        - name: discussionId
          in: path
          required: true
          style: simple
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/DiscussionDto'
            application/json:
              schema:
                $ref: '#/components/schemas/DiscussionDto'
            text/json:
              schema:
                $ref: '#/components/schemas/DiscussionDto'
  /api/v1/discussions:
    get:
      tags:
        - Discussions
      summary: Returns all discussions.
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DiscussionDto'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DiscussionDto'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DiscussionDto'
    post:
      tags:
        - Discussions
      summary: Creates a discussion.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiscussionRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/DiscussionRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/DiscussionRequest'
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/DiscussionDto'
            application/json:
              schema:
                $ref: '#/components/schemas/DiscussionDto'
            text/json:
              schema:
                $ref: '#/components/schemas/DiscussionDto'
  '/api/v1/discussions/{authorId}':
    get:
      tags:
        - Discussions
      summary: Returns all discussions created by author with the specified ID.
      parameters:
        - name: authorId
          in: query
          style: form
          schema:
            type: string
            format: uuid
        - name: authorId
          in: path
          required: true
          style: simple
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DiscussionDto'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DiscussionDto'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DiscussionDto'
components:
  schemas:
    AuthorDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        display_name:
          type: string
        avatar_url:
          type: string
          nullable: true
      additionalProperties: false
    CommentDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        content:
          type: string
        author:
          $ref: '#/components/schemas/AuthorDto'
        created_at:
          type: string
          format: date-time
        reply_count:
          type: integer
          format: int32
        first_replies_authors:
          type: array
          items:
            $ref: '#/components/schemas/AuthorDto'
      additionalProperties: false
    CommentRequest:
      type: object
      properties:
        discussion_id:
          type: string
          format: uuid
        content:
          type: string
      additionalProperties: false
    DiscussionDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        author:
          $ref: '#/components/schemas/AuthorDto'
      additionalProperties: false
    DiscussionRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
      additionalProperties: false
    ReplyRequest:
      type: object
      properties:
        content:
          type: string
      additionalProperties: false