# src/CrowdParlay.Social.Application/DTOs/DiscussionDto.cs
namespace CrowdParlay.Social.Application.DTOs;

public class DiscussionDto
{
    public Guid Id { get; set; }
    public string Title { get; set; }
    public string Description { get; set; }
    public AuthorDto Author { get; set; }
}